Phyton/requirements.txt
fastapi
uvicorn
mysql-connector-python
requests
python-crontab    # cron üçün
